+++ CHANGELOG.md
@@ -7,11 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 
+- OAuth2 client credentials rest template interceptor
 - Configuration for left and right policies to use registry client library
 
 ### Changed
 
+- Changed digitalTwinRegistryRestTemplate to use token in requests
 - Update asBuild test data to 1.5.3 and asPlanned to 1.5.1
+- Changed transfer notification logic not to break iteration loop when sending notifications to bpn with more than 1 connector endpoints
 
 ### Removed